pyproject.toml
[project]
name="SCP-obsidian"
version="0.0.1"
dependencies = [
    "bs4>=0.0.2",
    "libzim>=3.7.0",
    "mcp>=1.13.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "toml>=0.10.2",
]
[tool.Agent]
[tool.Agent.Dify]
base_url="localhost"
